+++ web_static/routes/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/python3
+
+from flask import Blueprint
+
+bp = Blueprint('views', __name__)
+
+from web_static.routes import views
