+++ web_static/templates/application_form.html
@@ -51,7 +51,7 @@
   <body>
     <div class="header-container">
       <nav>
-        <a href="{{ url_for('home') }}"
+        <a href="{{ url_for('views.home') }}"
           ><img
             src="{{ url_for('static', filename='images/careerlogo.png') }}"
             alt="logo"
@@ -73,11 +73,11 @@
             <line x1="6" y1="6" x2="18" y2="18" />
           </svg>
           <ul>
-            <li><a href="{{ url_for('home') }}">HOME</a></li>
-             <li><a href="{{ url_for('update_jobseeker_profile') }}">PROFILE</a></li>
-            <li><a href="{{ url_for('joblists') }}">JOBS</a></li>
-            <li><a href="{{ url_for('userDashboard') }}">DASHBOARD</a></li>
-            <a href="{{ url_for('logout') }}" class="hero-job">Log Out</a>
+            <li><a href="{{ url_for('views.home') }}">HOME</a></li>
+             <li><a href="{{ url_for('views.update_jobseeker_profile') }}">PROFILE</a></li>
+            <li><a href="{{ url_for('views.joblists') }}">JOBS</a></li>
+            <li><a href="{{ url_for('views.userDashboard') }}">DASHBOARD</a></li>
+            <a href="{{ url_for('views.logout') }}" class="hero-job">Log Out</a>
           </ul>
         </div>
         <svg
@@ -104,7 +104,7 @@
     <div class="form-container">
       <h2>Job Application Form</h2>
       <form
-        action="{{ url_for('applicationForm', id=job['id']) }}"
+        action="{{ url_for('views.applicationForm', id=job['id']) }}"
         enctype="multipart/form-data"
         method="post"
       >
