+++ web_static/templates/headerContent.html
@@ -1,74 +1,126 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>CareerLink</title>
-  <link rel="stylesheet" href="{{ url_for('static', filename='styles/style.css') }}">
-  <link rel="stylesheet" href="{{ url_for('static', filename='styles/footer.css') }}">
-  <link rel="stylesheet" href="{{ url_for('static', filename='styles/testimonials.css') }}">
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@fortawesome/fontawesome-free@6.2.1/css/fontawesome.min.css" integrity="sha384-QYIZto+st3yW+o8+5OHfT6S482Zsvz2WfOzpFSXMF9zqeLcFV0/wlZpMtyFcZALm" crossorigin="anonymous">
-  <!--font awesome link for icons-->
-  <!--font awesome link for icons must include this kit to prevent blank icons-->
-  <script src="https://kit.fontawesome.com/d7accf2a63.js" crossorigin="anonymous"></script>
-</head>
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>CareerLink</title>
+    <link
+      rel="stylesheet"
+      href="{{ url_for('static', filename='styles/style.css') }}"
+    />
+    <link
+      rel="stylesheet"
+      href="{{ url_for('static', filename='styles/footer.css') }}"
+    />
+    <link
+      rel="stylesheet"
+      href="{{ url_for('static', filename='styles/testimonials.css') }}"
+    />
+    <link
+      rel="stylesheet"
+      href="https://cdn.jsdelivr.net/npm/@fortawesome/fontawesome-free@6.2.1/css/fontawesome.min.css"
+      integrity="sha384-QYIZto+st3yW+o8+5OHfT6S482Zsvz2WfOzpFSXMF9zqeLcFV0/wlZpMtyFcZALm"
+      crossorigin="anonymous"
+    />
+    <!--font awesome link for icons-->
+    <!--font awesome link for icons must include this kit to prevent blank icons-->
+    <script
+      src="https://kit.fontawesome.com/d7accf2a63.js"
+      crossorigin="anonymous"
+    ></script>
+  </head>
 
-<body>
-
+  <body>
     <div class="header-container">
-    <nav>
-      <a href="{{ url_for('home') }}"><img src="{{ url_for('static', filename='images/careerlogo.png') }}" alt="logo"></a>
-      <div class="nav-links" id="navLinks">
-        <svg xmlns="http://www.w3.org/2000/svg" onclick="hideMenu()" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><line x1="18" y1="6" x2="6" y2="18"/><line x1="6" y1="6" x2="18" y2="18"/></svg>
-        <ul>
-          <li><a href="{{ url_for('home') }}">HOME</a></li>
-          <li><a href="{{ url_for('about') }}">ABOUT</a></li>
-          <li><a href="{{ url_for('joblists') }}">JOBS</a></li>
-          <li><a href="{{ url_for('contact') }}">CONTACT</a></li>
-          <a href="{{ url_for('login') }}" class="hero-job">Post a job</a>
-        </ul>
+      <nav>
+        <a href="{{ url_for('views.home') }}"
+          ><img
+            src="{{ url_for('static', filename='images/careerlogo.png') }}"
+            alt="logo"
+        /></a>
+        <div class="nav-links" id="navLinks">
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            onclick="hideMenu()"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+            stroke="currentColor"
+            stroke-width="2"
+            stroke-linecap="round"
+            stroke-linejoin="round"
+          >
+            <line x1="18" y1="6" x2="6" y2="18" />
+            <line x1="6" y1="6" x2="18" y2="18" />
+          </svg>
+          <ul>
+            <li><a href="{{ url_for('views.home') }}">HOME</a></li>
+            <li><a href="{{ url_for('views.about') }}">ABOUT</a></li>
+            <li><a href="{{ url_for('views.joblists') }}">JOBS</a></li>
+            <li><a href="{{ url_for('views.contact') }}">CONTACT</a></li>
+            <a href="{{ url_for('views.login') }}" class="hero-job"
+              >Post a job</a
+            >
+          </ul>
+        </div>
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          onclick="showMenu()"
+          width="24"
+          height="24"
+          viewBox="0 0 24 24"
+          fill="none"
+          stroke="currentColor"
+          stroke-width="2"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+        >
+          <line x1="3" y1="12" x2="21" y2="12" />
+          <line x1="3" y1="6" x2="21" y2="6" />
+          <line x1="3" y1="18" x2="21" y2="18" />
+        </svg>
+      </nav>
+    </div>
+
+    <!--footer-->
+    <footer class="footer">
+      <div class="container">
+        <div class="row">
+          <div class="footer-col">
+            <h4>company</h4>
+            <ul>
+              <li><a href="#">Cookie Policy</a></li>
+              <li><a href="#">Terms of Use</a></li>
+              <li><a href="#">privacy policy</a></li>
+              <li><a href="#">affiliate program</a></li>
+              <li><a href="#">Accessibility</a></li>
+            </ul>
+          </div>
+          <div class="footer-col">
+            <h4>get help</h4>
+            <ul>
+              <li><a href="#">FAQ</a></li>
+              <li><a href="#">Terms and Conditions</a></li>
+              <li><a href="#">locate a service provider</a></li>
+              <li><a href="#">Help</a></li>
+              <li><a href="#">Report a problem</a></li>
+            </ul>
+          </div>
+
+          <div class="footer-col">
+            <h4>follow us</h4>
+            <div class="social-links">
+              <a href="#"><i class="fa-brands fa-facebook"></i></a>
+              <a href="#"><i class="fa-brands fa-instagram"></i></a>
+              <a href="#"><i class="fa-brands fa-twitter"></i></a>
+              <a href="#"><i class="fa-brands fa-pinterest"></i></a>
+            </div>
+          </div>
+        </div>
       </div>
-      <svg xmlns="http://www.w3.org/2000/svg" onclick="showMenu()" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><line x1="3" y1="12" x2="21" y2="12"/><line x1="3" y1="6" x2="21" y2="6"/><line x1="3" y1="18" x2="21" y2="18"/></svg>
-    </nav>
-  </div>
-
-
-  <!--footer-->
-  <footer class="footer">
-  	 <div class="container">
-  	 	<div class="row">
-  	 		<div class="footer-col">
-  	 			<h4>company</h4>
-  	 			<ul>
-  	 				<li><a href="#">Cookie Policy</a></li>
-  	 				<li><a href="#">Terms of Use</a></li>
-  	 				<li><a href="#">privacy policy</a></li>
-  	 				<li><a href="#">affiliate program</a></li>
-            <li><a href="#">Accessibility</a></li>
-  	 			</ul>
-  	 		</div>
-  	 		<div class="footer-col">
-  	 			<h4>get help</h4>
-  	 			<ul>
-  	 				<li><a href="#">FAQ</a></li>
-  	 				<li><a href="#">Terms and Conditions</a></li>
-  	 				<li><a href="#">locate a service provider</a></li>
-  	 				<li><a href="#">Help</a></li>
-  	 				<li><a href="#">Report a problem</a></li>
-  	 			</ul>
-  	 		</div>
-
-  	 		<div class="footer-col">
-  	 			<h4>follow us</h4>
-  	 			<div class="social-links">
-  	 				<a href="#"><i class="fa-brands fa-facebook"></i></a>
-  	 				<a href="#"><i class="fa-brands fa-instagram"></i></a>
-  	 				<a href="#"><i class="fa-brands fa-twitter"></i></a>
-  	 				<a href="#"><i class="fa-brands fa-pinterest"></i></a>
-  	 			</div>
-  	 		</div>
-  	 	</div>
-  	 </div>
-     <hr>
-    <p class="copyright">© CareerLink 2023. All rights reserved</p>
-  </footer>
+      <hr />
+      <p class="copyright">© CareerLink 2023. All rights reserved</p>
+    </footer>
+  </body>
+</html>
