+++ web_static/templates/joblists.html
@@ -73,7 +73,7 @@
     <!-- Header Section -->
     <div class="header-container">
       <nav>
-        <a href="{{ url_for('home') }}"
+        <a href="{{ url_for('views.home') }}"
           ><img
             src="{{ url_for('static', filename='images/careerlogo.png') }}"
             alt="logo"
@@ -95,11 +95,19 @@
             <line x1="6" y1="6" x2="18" y2="18" />
           </svg>
           <ul>
-            <li><a href="{{ url_for('home') }}">HOME</a></li>
-            <li><a href="{{ url_for('update_jobseeker_profile') }}">PROFILE</a></li>
-            <li><a href="{{ url_for('userDashboard') }}">DASHBOARD</a></li>
-            <li><a href="{{ url_for('my_applied_jobs') }}">APPLIED JOBS</a></li>
-            <a href="{{ url_for('logout') }}" class="hero-job">Log Out</a>
+            <li><a href="{{ url_for('views.home') }}">HOME</a></li>
+            <li>
+              <a href="{{ url_for('views.update_jobseeker_profile') }}"
+                >PROFILE</a
+              >
+            </li>
+            <li>
+              <a href="{{ url_for('views.userDashboard') }}">DASHBOARD</a>
+            </li>
+            <li>
+              <a href="{{ url_for('views.my_applied_jobs') }}">APPLIED JOBS</a>
+            </li>
+            <a href="{{ url_for('views.logout') }}" class="hero-job">Log Out</a>
           </ul>
         </div>
         <svg
